+++ GameCoreLibrary/Classes/ObjectStats.cs
@@ -4,11 +4,11 @@ namespace GameCoreLibrary.Classes
 {
     public abstract class ObjectStats
     {
-        public Dictionary<string, int> Stats { get; set; }
+        public Dictionary<string, double> Stats { get; set; }
 
-        protected ObjectStats(Dictionary<string, int> stats)
+        protected ObjectStats(Dictionary<string, double> stats)
         {
-            Stats = new Dictionary<string, int>(stats);
+            Stats = new Dictionary<string, double>(stats);
         }
 
     }
